Use regex to determine if Csound test passed
Csound sets its exit code to the number of init errors, so it can't be used to determine if the test passed or not if init errors are expected.

This change fixes the issue by using the presence of "All tests passed." to determing if the test passed or not instead of Csound's exit code.